acitivities and import updates
